requirements.txt
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
configparser==7.0.0
Flask==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.2.2
MarkupSafe==2.1.5
pysafebrowsing==0.1.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-whois==0.9.4
requests==2.32.3
six==1.16.0
soupsieve==2.5
threatminer==1.0
tld==0.13
urllib3==2.2.2
validators==0.33.0
virustotal-python==1.0.2
waybackpy==3.0.6
Werkzeug==3.0.3
